+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Release 1.5.0
+
+## API
+
+* New carpool proof system :
+    * Automatic creation of carpool proof for planified carpools
+    * Realtime creation of carpool proofs for planified and dynamic carpools using mobile app
+    * Link with proof registry via cron job
+* New entities for SolidaryUser : SolidaryBeneficiary and SolidaryVolunteer.
+* Several utility routes.
+
 # Release 1.4.0
 
 ## API